This is a Dockerfile used to create a Docker image for a Node.js application. The image will be based on the Alpine Linux distribution.
The maintainer of the image is Jonathan Baraldi.

The image will include a /src directory, which will be the working directory for the application. All files in the current directory will be added to this directory in the image.

The application will be accessible on port 8080.

Finally, the image will run the Node.js application located in /src/index.js when started.