# ChatGPT CodeReview This GitHub Action will use the ChatGPT OpenAI engine to review pull request events when opened or synchronized.
## Job
The job will run on an Ubuntu-latest runner and will be named ChatGPT explain code.

## Steps
The step will use the ChatGPT Github Actions from the cirolini repository. It will also get the OpenAI API key and Github Token from the repository secrets. It will also get the Github Pull Request ID from the Github event. Additionally, an optional OpenAI engine and sampling temperature can be specified. The default OpenAI engine is "text-davinci-002" and the default sampling temperature is 0.5. The default maximum number of tokens to generate in OpenAI completion is 2048.