#
This Dockerfile will create an image based on the Alpine Linux distribution of Node.js. The `RUN` command will create a directory called `/src` and the `WORKDIR` command will set the working directory to `/src`. The `ADD` command will add all files from the current directory to the `/src` directory. The `EXPOSE` command will expose port 8080. The `CMD` command will execute the `index.js` file in the `/src` directory when the image is run.