# ChatGPT CodeReview This GitHub Action runs code reviews with ChatGPT. It is triggered on pull request events when they are opened or synchronized.
# hello_world_job
This job runs on an ubuntu-latest runner and its name is ChatGPT explain code.

# Step
The action uses the ChatGPT Github Actions from cirolini repository. It gets the OpenAI API key and Github Token from repository secrets and the Github Pull Request ID from the Github event. Additionally, it allows to specify the OpenAI engine to use (default is "text-davinci-002") and the sampling temperature for OpenAI (default is 0.5) and the maximum number of tokens to generate in OpenAI completion (default is 2048).