# ChatGPT CodeReview This GitHub Action runs a code review using ChatGPT. It is triggered on pull request events when opened or synchronized.
The job `hello_world_job` is defined and is run on the `ubuntu-latest` runner. It uses the ChatGPT Github Actions from the `jonathanbaraldi/github-actions-chatgpt@main` repository and requires the OpenAI API key and Github Token from the repository secrets. The Github Pull Request ID is also taken from the Github event. 

Optional parameters can be specified for the OpenAI engine, sampling temperature, and maximum number of tokens to generate in OpenAI completion. The defaults are "text-davinci-002" for the engine, 0.5 for the temperature, and 2048 for the maximum number of tokens.