# ChatGPT CodeReview This GitHub Action will use ChatGPT to explain code in a Pull Request.
## Trigger the workflow
The workflow will be triggered on pull request events when the pull request is opened or synchronized.

## Job
The job will be run on an Ubuntu-latest runner and will be named "ChatGPT explain code".

## Step
The step will use the ChatGPT GitHub Actions from cirolini repository and will get the OpenAI API key and Github Token from the repository secrets. The Github Pull Request ID will be retrieved from the Github event and the OpenAI engine, temperature and maximum number of tokens to generate in OpenAI completion can be specified (defaults are "text-davinci-002", 0.5 and 2048 respectively).