# ChatGPT CodeReview This GitHub Action runs ChatGPT to review pull requests on your repository.
It is triggered on pull request events when the pull request is opened or synchronized.

It runs the job on an Ubuntu-latest runner, and uses the ChatGPT Github Actions from the cirolini repository. It gets the OpenAI API key and Github Token from repository secrets, and the Github Pull Request ID from the Github event. 

The OpenAI engine and sampling temperature for OpenAI can be specified, as well as the maximum number of tokens to generate in OpenAI completion.