# ChatGPT CodeReview This GitHub Action is designed to run a code review using ChatGPT. It is triggered on pull request events when opened or synchronized.
# Job
The job is defined as `hello_world_job` and runs on the `ubuntu-latest` runner.

# Step
The step is named `ChatGPT explain code` and uses the ChatGPT Github Actions from cirolini repository. The OpenAI API key and Github Token are retrieved from the repository secrets. The Github Pull Request ID is retrieved from the Github event. The OpenAI engine, sampling temperature, and maximum number of tokens to generate in OpenAI completion are all optional and have default values.