# ChatGPT CodeReview This GitHub Action is used to review pull requests using ChatGPT. It will use the OpenAI API to generate comments for the pull request based on the code changes.
# on
This section specifies when the workflow should be triggered. In this case, the workflow will be triggered when a pull request is opened or synchronized.

# jobs
This section defines the jobs that will be run by the workflow. In this case, the job is called `hello_world_job` and it will be run on an Ubuntu-latest runner.

# steps
This section defines the steps that will be run by the job. In this case, the step is called `ChatGPT explain code` and it will use the ChatGPT Github Actions from the cirolini repository. It will also use the OpenAI API key, Github Token, and Github Pull Request ID from the repository secrets, as well as the OpenAI engine, sampling temperature, and maximum number of tokens to generate in OpenAI completion.