# ChatGPT CodeReview This GitHub Action runs the ChatGPT Github Action from the cirolini repository when a pull request is opened or synchronized.
The job is named `ChatGPT explain code` and it runs on an `ubuntu-latest` runner. 

The required parameters for the ChatGPT Github Action are the `openai_api_key`, `github_token` and `github_pr_id`. Additionally, the `openai_engine`, `openai_temperature` and `openai_max_tokens` parameters can be specified for further customization. These parameters are all sourced from the repository secrets.