# ChatGPT CodeReview This GitHub Action runs ChatGPT to explain code in Pull Requests.
It is triggered when a pull request is opened or synchronized and runs on a Ubuntu-latest runner.

The Action uses the ChatGPT Github Actions from cirolini repository and requires an OpenAI API key and Github Token from repository secrets. It also takes the Github Pull Request ID from the Github event and allows for optional parameters such as the OpenAI engine, sampling temperature, and maximum number of tokens to generate.