# ChatGPT CodeReview This workflow is triggered when a pull request is opened or synchronized and is used to run the ChatGPT Github Actions from the cirolini repository.
# Job
The job is defined as `hello_world_job` and runs on the `ubuntu-latest` runner.

# Step
The step is named `ChatGPT explain code` and uses the ChatGPT Github Actions from the cirolini repository. It utilizes the OpenAI API key from repository secrets, the Github Token from repository secrets, the Github Pull Request ID from the Github event, and optionally, the OpenAI engine, sampling temperature, and maximum number of tokens to generate in OpenAI completion.