# ChatGPT CodeReview This GitHub Action will use the ChatGPT Github Actions from the cirolini repository to explain code from a pull request.
# On
This workflow will be triggered on pull request events when opened or synchronized.

# Jobs
This job will run on Ubuntu-latest and the name of the job is ChatGPT explain code.

# Steps
This step will use the ChatGPT Github Actions from cirolini repository. The OpenAI API key and Github Token will be retrieved from repository secrets. The Github Pull Request ID will be retrieved from the Github event. The OpenAI engine, sampling temperature, and maximum number of tokens to generate in OpenAI completion can also be specified.