# ChatGPT CodeReview This GitHub Action uses [jonathanbaraldi/github-actions-chatgpt@main](https://github.com/jonathanbaraldi/github-actions-chatgpt) to run a ChatGPT explain code job on pull request events when the pull request is opened or synchronized.
The job requires an OpenAI API key and a GitHub token, both of which are retrieved from the repository secrets. The GitHub Pull Request ID is retrieved from the GitHub event.

Optional parameters can be specified to customize the OpenAI engine, sampling temperature, and maximum number of tokens to generate in OpenAI completion. If no parameters are specified, the defaults are "text-davinci-002", 0.5, and 2048 respectively.