# ChatGPT CodeReview This GitHub Action runs a code review using ChatGPT by OpenAI. It is triggered on pull request events when they are opened or synchronized.
The job runs on an Ubuntu-latest runner and has the name "ChatGPT explain code". The action uses the ChatGPT Github Actions from the cirolini repository and requires an OpenAI API key and Github Token from the repository secrets. The Github Pull Request ID is obtained from the Github event. 

The OpenAI engine and sampling temperature can be specified as optional parameters. The maximum number of tokens to generate in OpenAI completion is set to 2048 by default.