# ChatGPT CodeReview This GitHub Action is used to review code using ChatGPT. It will trigger on pull request events when opened or synchronized.
The job `hello_world_job` is defined and runs on `ubuntu-latest`. It will use the ChatGPT Github Actions from cirolini repository and get the OpenAI API key and Github Token from repository secrets. It will also get the Github Pull Request ID from the Github event. 

The OpenAI engine, sampling temperature, and maximum number of tokens to generate in OpenAI completion can also be specified as optional parameters. The default values are "text-davinci-002" for the engine, 0.5 for the sampling temperature, and 2048 for the maximum number of tokens.