# ChatGPT CodeReview This GitHub Action will trigger ChatGPT to explain code in a pull request.
# on: pull_request
This event listener will trigger the workflow on pull request events when the pull request is opened or synchronized.

# jobs: hello_world_job
This job will run on an ubuntu-latest runner, and will be given the name ChatGPT explain code.

# steps:
This step will use the ChatGPT Github Actions from cirolini repository. It will use the OpenAI API key and Github Token from repository secrets, as well as the Github Pull Request ID from the Github event. It will also optionally specify the OpenAI engine, sampling temperature, and maximum number of tokens to generate in OpenAI completion.